Keep old slot migration behaviors (#65)
when performing atomic slot migration, during key importing on the destination node or key trimming on the source/destination, these keys will be filtered out in the following commands:
- KEYS
- SCAN
- RANDOMKEY
- CLUSTER GETKEYSINSLOT
- DBSIZE
- CLUSTER COUNTKEYSINSLOT

The only command that will reflect the increasing number of keys is:
- INFO KEYSPACE

### Downsides
- Replica behavior
For replicas, importing keys will also be filtered if using the old slot migration approach. This means the above commands may show breaking changes: in the previous version they were visible, but in the new version they are not. The reason is that replicas cannot know which slots are being imported or migrated.
- DBSIZE performance
The DBSIZE command may experience performance regression, since it now requires iterating over all slot dicts to determine whether to include keys in the slot.